backend/requirements.txt
asgiref==3.8.1
click==8.1.7
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
h11==0.14.0
sqlparse==0.5.2
typing_extensions==4.12.2
uvicorn==0.32.1
